Integrated Diffie Hellman into SSL handshake. Also moved SSL Handshake to it's own folder.
