+++ Functions.py
@@ -15,6 +15,7 @@ from cryptography.hazmat.primitives import padding as sym_padding
 from cryptography.hazmat.primitives.asymmetric import rsa, padding, dh
 from cryptography.hazmat.primitives.ciphers import algorithms, Cipher, modes
 from cryptography.hazmat.primitives.kdf.hkdf import HKDF
+from cryptography.hazmat.primitives.serialization import Encoding, PublicFormat, load_der_public_key
 from cryptography.x509.oid import NameOID
 
 from Constants import p, g
@@ -150,6 +151,27 @@ def rsa_public_encrypt(public_key, msg_in_bytes):
     )
     return ciphertext
 
+def rsa_private_sign(privit_key, msg_in_bytes):
+    signed_text = privit_key.sign(
+        msg_in_bytes,
+        padding.PSS(
+            mgf=padding.MGF1(hashes.SHA256()),
+            salt_length=padding.PSS.MAX_LENGTH
+        ),
+        hashes.SHA256()
+    )
+    return signed_text
+
+def rsa_signature_verification(public_key, signed_msg, unsigned_msg):
+    public_key.verify(
+        signed_msg,
+        unsigned_msg,
+        padding.PSS(
+            mgf=padding.MGF1(hashes.SHA256()),
+            salt_length=padding.PSS.MAX_LENGTH
+        ),
+        hashes.SHA256()
+    )
 
 def rsa_private_decrypt(private_key, ciphertext):
     decrypted_text = private_key.decrypt(
@@ -197,145 +219,6 @@ def generate4keys(master_key):
             get_SHA_1_Hash(key_input2),
             get_SHA_1_Hash(key_input3),
             get_SHA_1_Hash(master_key)]
-
-
-# --------------------- Modified SSL Handshake -----------------------
-def server_ssl_handshake(node_socket, node_name, private_key):
-    all_msgs = ""
-
-    # ---- msg 1
-    received_msg1 = read_message_with_delimiter(node_socket)
-    all_msgs += str(received_msg1)
-    print(f"[{node_name}] Received: ", str(received_msg1))
-
-    # determine encryption/integrity algorithms
-    if "AES" in received_msg1["supported_data_encryption_algs"]:
-        print(f"[{node_name}] Chosen Encryption is AES")
-    else:
-        print(f"[{node_name}] Unsupported data encryption algs")
-        node_socket.close()
-        return False
-    if "AES" in received_msg1["supported_integrity_algs"]:
-        print(f"[{node_name}] Chosen Integrity protection is AES")
-    else:
-        print(f"[{node_name}] Unsupported integrity algs")
-        node_socket.close()
-        return False
-
-    online_node_certificate = loadCert(received_msg1["certificate"])
-    online_node_public_key = online_node_certificate.public_key()
-
-    # verify signature
-    try:
-        verifyCertificateSignature(online_node_certificate, online_node_public_key)
-    except cryptography.exceptions.InvalidSignature:
-        print(f"[{node_name}] Signaturew was not valid.")
-        node_socket.close()
-        return False
-
-    # create certificate
-    cert = createCertificate(private_key)
-    # ---- msg 2
-    msg2 = {"certificate": certificate_to_byes(cert),
-            "chosen_data_encryption_alg": "AES",
-            "chosen_integrity_alg": "AES"}
-    all_msgs += str(msg2)
-    print(f"[{node_name}] Sending: ", msg2)
-    node_socket.sendall(wrap_to_send(msg2))
-
-    print(f"First part of Handshake Working!")
-
-    # ---- DO DH for creating keys
-
-    # Generate MAC
-    created_MAC = create_MAC(all_msgs)
-    msg3_5 = {"mac": created_MAC}
-
-    all_msgs += str(msg3_5)
-    print(f"[{node_name}] Sending: ", str(msg3_5))
-    node_socket.sendall(wrap_to_send(msg3_5))
-
-    # Receive MAC
-    msg4 = read_message_with_delimiter(node_socket)
-    print(f"[{node_name}] Received: ", str(msg4))
-    all_msgs_before_mac = all_msgs
-    all_msgs += str(msg4)
-    online_node_mac = msg4["mac"]
-
-    # verify MAC
-    if valid_mac(all_msgs_before_mac, online_node_mac):
-        print(f"[{node_name}] MAC is Valid")
-    else:
-        print(f"[{node_name}] MAC was not validated.")
-        node_socket.close()
-        return False
-
-    return True
-
-
-def client_ssl_handshake(node_socket, node_name, private_key):
-    all_msgs = ""
-
-    # ---- msg 1
-    supported_data_encryption_algs = ["AES"]
-    supported_integrity_algs = ["AES"]
-
-    cert = createCertificate(private_key)
-
-    msg1 = {"certificate": certificate_to_byes(cert),
-            "supported_data_encryption_algs": supported_data_encryption_algs,
-            "supported_integrity_algs": supported_integrity_algs}
-
-    all_msgs += str(msg1)
-    print("[Online Node] Sending: ", msg1)
-    node_socket.sendall(wrap_to_send(msg1))
-
-    # ---- msg 2
-    received_msg2 = read_message_with_delimiter(node_socket)
-    print("[Online Node] Received: ", str(received_msg2))
-    all_msgs += str(received_msg2)
-
-    middle_node_certificate = loadCert(received_msg2["certificate"])
-    middle_node_public_key = middle_node_certificate.public_key()
-
-    # verify certificate
-    print("[Online Node] Verifying Certificate... ", end="")
-    try:
-        verifyCertificateSignature(middle_node_certificate, middle_node_public_key)
-    except cryptography.exceptions.InvalidSignature:
-        print("Verification Failed!")
-        node_socket.close()
-        return False
-
-    print("Certificate Verification Complete!")
-
-    # Receive MAC
-    msg4 = read_message_with_delimiter(node_socket)
-    print("[Middle Node] Received: ", str(msg4))
-    all_msgs_before_mac = all_msgs
-    all_msgs += str(msg4)
-    online_node_mac = msg4["mac"]
-
-    # verify MAC
-    if valid_mac(all_msgs_before_mac, online_node_mac):
-        print("[Middle Node] MAC is Valid")
-    else:
-        print("[Middle Node] MAC was not validated.")
-        node_socket.close()
-        return False
-
-    # Generate MAC
-    created_MAC = create_MAC(all_msgs)
-    msg3_5 = {"mac": created_MAC}
-
-    all_msgs_before_mac = all_msgs
-    all_msgs += str(msg3_5)
-    print("[Middle Node] Sending: ", str(msg3_5))
-    node_socket.sendall(wrap_to_send(msg3_5))
-
-    print("FINISHED!")
-    node_socket.close()
-    return True
 
 
 # --------- Data Transfer Phase ----------------
